Since jikes 1.07 (the introduction of the unified block stack), jikes has segfaulted when compiling certain switch-statement constructs.  After grovelling through the code for far too long, I've finally figured out why, and created the attached patch (against CVS-current) to fix the problem.
First, the why:  buried way down deep in the parser code, somebody
double-converted to a TokenIndex.  First some function explanations:
the expression Token(i) gets the TokenIndex corresponding to the i'th
element in the parser production rule, and Sym(i) gets the AST subtree
associated with the i'th element in the parser production rule.  The
expression Sym(i)->RightToken() gets the TokenIndex associated with the
right-most token in the AST subtree associated with the i'th element in
the parser production rule.  Clear as mud?

The NewEmptyStatement(token) constructor takes a TokenIndex.

In line 2975 of java.g, the constructor expression *should* read:
  NewEmptyStatement(Sym(3) -> RightToken())
but someone went Token()-happy and wrote
  NewEmptyStatement(Token(Sym(3) -> RightToken()))
instead.  Now tokens are large-ish integers, and this particular error
caused an evaluation of the expression Token(542)... for a grammar
production with only 4 elements.  End result: garbage in the token
field of the AstEmptyStatement created.  Which doesn't cause problems
until you eventually try to either:
a) print out the AST (which causes predictable death when it trys to
   look up the garbage token), or
b) compile to bytecode, which segfaults trying to look up a line number
   for the garbage token in order to add an entry to the line number
   table attribute.
Both quite cute.

The attached patch fixes the problem.  Only the change to java.g is
strictly needed, but to save the maintainers the trouble of running
javapg again, I've done it for y'all and included the appropriate patch
to the generated javaact.cpp file as well.
  --s

Obtained from: C. Scott Ananian <cananian@lesser-magoo.lcs.mit.edu>
Submitted by: C. Scott Ananian <cananian@lesser-magoo.lcs.mit.edu>
Reviewed by: Vadim Zaliva <lord@crocodile.org>